style: 기존 디렉토리 구조 변경
도메인별 최상위 디렉토리를 생성하는 방식에서 최상위 디렉토리에 model을 두고 하위 디렉토리로 생성하는 방식으로 변경
