+++ src/main/java/com/compono/ibackend/auth/service/AuthService.java
@@ -4,8 +4,8 @@ import com.compono.ibackend.auth.dto.response.AuthRefreshResponse;
 import com.compono.ibackend.common.enumType.ErrorCode;
 import com.compono.ibackend.common.exception.CustomException;
 import com.compono.ibackend.common.utils.jwt.JwtProvider;
-import com.compono.ibackend.user.domain.User;
-import com.compono.ibackend.user.repository.UserRepository;
+import com.compono.ibackend.model.user.domain.User;
+import com.compono.ibackend.model.user.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
