+++ src/main/java/com/compono/ibackend/model/schedule/dto/ScheduleTimeDTO.java
@@ -1,0 +1,12 @@
+package com.compono.ibackend.model.schedule.dto;
+
+import com.compono.ibackend.model.scheduleTime.domain.ScheduleTime;
+import java.time.LocalDateTime;
+
+public record ScheduleTimeDTO(Long id, LocalDateTime stopTime, LocalDateTime startTime) {
+
+    public static ScheduleTimeDTO from(ScheduleTime scheduleTime) {
+        return new ScheduleTimeDTO(
+                scheduleTime.getId(), scheduleTime.getStopTime(), scheduleTime.getStartTime());
+    }
+}
