+++ src/main/java/com/compono/ibackend/model/schedule/dto/request/TimeLineRequest.java
@@ -1,0 +1,5 @@
+package com.compono.ibackend.model.schedule.dto.request;
+
+import java.time.LocalDateTime;
+
+public record TimeLineRequest(Long userId, LocalDateTime specificTime) {}
