+++ src/main/java/com/compono/ibackend/model/schedule/dto/response/TagDetailResponse.java
@@ -1,0 +1,3 @@
+package com.compono.ibackend.model.schedule.dto.response;
+
+public record TagDetailResponse(Long tagId, String name, String color) {}
