+++ src/main/java/com/compono/ibackend/model/schedule/enumType/SchedulePriority.java
@@ -1,0 +1,21 @@
+package com.compono.ibackend.model.schedule.enumType;
+
+import java.util.Arrays;
+import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum SchedulePriority {
+    P1(0, "상"),
+    P2(1, "중"),
+    P3(2, "하");
+
+    private final int code;
+    private final String value;
+
+    public static String getAllType() {
+        return Arrays.stream(values())
+                .map(SchedulePriority::name)
+                .collect(Collectors.joining(" || "));
+    }
+}
