+++ src/main/java/com/compono/ibackend/model/schedule/enumType/TaskStatus.java
@@ -1,0 +1,14 @@
+package com.compono.ibackend.model.schedule.enumType;
+
+import java.util.Arrays;
+import java.util.stream.Collectors;
+
+public enum TaskStatus {
+    IN_PROGRESS,
+    PAUSED,
+    COMPLETED;
+
+    public static String getAllType() {
+        return Arrays.stream(values()).map(TaskStatus::name).collect(Collectors.joining(" || "));
+    }
+}
