+++ src/main/java/com/compono/ibackend/model/schedule/repository/ScheduleTimeRepository.java
@@ -1,0 +1,6 @@
+package com.compono.ibackend.model.schedule.repository;
+
+import com.compono.ibackend.model.scheduleTime.domain.ScheduleTime;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ScheduleTimeRepository extends JpaRepository<ScheduleTime, Long> {}
