+++ src/main/java/com/compono/ibackend/model/user/enumType/OauthProvider.java
@@ -1,0 +1,14 @@
+package com.compono.ibackend.model.user.enumType;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum OauthProvider {
+    GOOGLE("google"),
+    KAKAO("kakao"),
+    APPLE("apple");
+
+    private final String value;
+}
