+++ src/main/java/com/compono/ibackend/model/user/enumType/UserStatus.java
@@ -1,0 +1,7 @@
+package com.compono.ibackend.model.user.enumType;
+
+public enum UserStatus {
+    ACTIVE,
+    DORMANT,
+    DELETED
+}
