+++ src/main/java/com/compono/ibackend/oauth/service/KakaoOauthService.java
@@ -7,9 +7,9 @@ import com.compono.ibackend.oauth.domain.KakaoUserInfo;
 import com.compono.ibackend.oauth.dto.OauthTokenDTO;
 import com.compono.ibackend.oauth.dto.request.OauthLoginRequest;
 import com.compono.ibackend.oauth.dto.response.OauthLoginResponse;
-import com.compono.ibackend.user.domain.User;
-import com.compono.ibackend.user.enumType.OauthProvider;
-import com.compono.ibackend.user.repository.UserRepository;
+import com.compono.ibackend.model.user.domain.User;
+import com.compono.ibackend.model.user.enumType.OauthProvider;
+import com.compono.ibackend.model.user.repository.UserRepository;
 import jakarta.servlet.http.HttpServletResponse;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
