--- src/main/java/com/compono/ibackend/schedule/dto/request/TimeLineRequest.java
@@ -1,5 +1,0 @@
-package com.compono.ibackend.schedule.dto.request;
-
-import java.time.LocalDateTime;
-
-public record TimeLineRequest(Long userId, LocalDateTime specificTime) {}
