--- src/main/java/com/compono/ibackend/schedule/dto/response/ScheduleTimeResponse.java
@@ -1,12 +1,0 @@
-package com.compono.ibackend.schedule.dto.response;
-
-import com.compono.ibackend.schedule.domain.ScheduleTime;
-import java.time.LocalDateTime;
-
-public record ScheduleTimeResponse(Long id, LocalDateTime startTime, LocalDateTime stopTime) {
-
-    public static ScheduleTimeResponse from(ScheduleTime scheduleTime) {
-        return new ScheduleTimeResponse(
-                scheduleTime.getId(), scheduleTime.getStartTime(), scheduleTime.getStopTime());
-    }
-}
