--- src/main/java/com/compono/ibackend/schedule/dto/response/TagDetailResponse.java
@@ -1,3 +1,0 @@
-package com.compono.ibackend.schedule.dto.response;
-
-public record TagDetailResponse(Long tagId, String name, String color) {}
