--- src/main/java/com/compono/ibackend/user/enumType/UserStatus.java
@@ -1,7 +1,0 @@
-package com.compono.ibackend.user.enumType;
-
-public enum UserStatus {
-    ACTIVE,
-    DORMANT,
-    DELETED
-}
