+++ src/main/java/com/compono/ibackend/web/user/UserController.java
@@ -1,0 +1,26 @@
+package com.compono.ibackend.web.user;
+
+import com.compono.ibackend.model.user.dto.request.UserAddRequest;
+import com.compono.ibackend.model.user.dto.response.UserAddResponse;
+import com.compono.ibackend.service.user.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/users")
+public class UserController {
+
+    private final UserService userService;
+
+    @PostMapping
+    public ResponseEntity<UserAddResponse> addUser(@RequestBody UserAddRequest request) {
+        UserAddResponse response = userService.addUser(request);
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    }
+}
