+++ src/test/java/com/compono/ibackend/auth/service/AuthServiceTest.java
@@ -8,10 +8,10 @@ import static org.mockito.Mockito.when;
 import com.compono.ibackend.auth.dto.response.AuthRefreshResponse;
 import com.compono.ibackend.common.exception.CustomException;
 import com.compono.ibackend.common.utils.jwt.JwtProvider;
-import com.compono.ibackend.user.domain.User;
-import com.compono.ibackend.user.dto.request.UserAddRequest;
-import com.compono.ibackend.user.enumType.OauthProvider;
-import com.compono.ibackend.user.repository.UserRepository;
+import com.compono.ibackend.model.user.domain.User;
+import com.compono.ibackend.model.user.dto.request.UserAddRequest;
+import com.compono.ibackend.model.user.enumType.OauthProvider;
+import com.compono.ibackend.model.user.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import jakarta.servlet.http.Cookie;
