+++ src/test/java/com/compono/ibackend/common/builder/BuilderSupporter.java
@@ -1,8 +1,7 @@
 package com.compono.ibackend.common.builder;
 
-import com.compono.ibackend.schedule.repository.ScheduleRepository;
-import com.compono.ibackend.schedule.repository.ScheduleTimeRepository;
-import com.compono.ibackend.user.repository.UserRepository;
+import com.compono.ibackend.model.schedule.repository.ScheduleTimeRepository;
+import com.compono.ibackend.model.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -11,16 +10,10 @@ public class BuilderSupporter {
 
     @Autowired private ScheduleTimeRepository scheduleTimeRepository;
 
-    @Autowired private ScheduleRepository scheduleRepository;
-
     @Autowired private UserRepository userRepository;
 
     public ScheduleTimeRepository scheduleTimeRepository() {
         return scheduleTimeRepository;
-    }
-
-    public ScheduleRepository scheduleRepository() {
-        return scheduleRepository;
     }
 
     public UserRepository userRepository() {
